Update StreamLogsCard tests after refactor
